style_parser: parser: Include span nigra in small red text fragment
For some rubrics text, there is a span with class "nigra" used inside
font element. It allows to get more detailed information from the text line.
However, as this information is not yet required, just add a text from the span
to regular small text fragment for simplicity.

This change fixes parsing of offices that includes the span nigra element.
Previously, parsing was failing in such case.

Improve logging when already on this.